Added a button to set the engine primitives to the current selection.
git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@1045 af78ad9f-9711-0410-ba57-a8eff4914ec4

+++ avogadro/src/extensions/selectextension.cpp
@@ -181,7 +181,7 @@ namespace Avogadro {
       return;
     */
     FOR_ATOMS_OF_MOL (atom, m_molecule) {
-      if (atom->GetAtomicNum() == element)
+      if (atom->GetAtomicNum() == static_cast<unsigned int>(element))
         selectedAtoms.append(static_cast<Atom*>(m_molecule->GetAtom(atom->GetIdx())));
     }
